requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
et_xmlfile==2.0.0
fastapi==0.123.0
h11==0.16.0
idna==3.11
numpy==2.3.5
openpyxl==3.1.5
pandas==2.3.3
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
six==1.17.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
xlrd==2.0.2
